+++ server/src/main/java/com/main19/server/chatroom/controller/ChatRoomController.java
@@ -1,25 +1,21 @@
 package com.main19.server.chatroom.controller;
 
-import com.main19.server.auth.Login;
+import com.main19.server.utils.Login;
 import com.main19.server.chatroom.dto.ChatRoomDto;
 import com.main19.server.chatroom.entity.ChatRoom;
 import com.main19.server.chatroom.mapper.ChatRoomMapper;
 import com.main19.server.chatroom.service.ChatRoomService;
 import com.main19.server.dto.SingleResponseDto;
-import com.main19.server.exception.BusinessLogicException;
-import com.main19.server.exception.ExceptionCode;
 import com.main19.server.member.entity.Member;
 import java.util.List;
 import javax.validation.constraints.Positive;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.MissingRequestHeaderException;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
