+++ server/src/main/java/com/main19/server/follow/controller/FollowController.java
@@ -1,9 +1,7 @@
 package com.main19.server.follow.controller;
 
-import com.main19.server.auth.Login;
-import com.main19.server.auth.jwt.JwtTokenizer;
+import com.main19.server.utils.Login;
 import com.main19.server.dto.SingleResponseDto;
-import com.main19.server.follow.dto.FollowDto;
 import com.main19.server.follow.entity.Follow;
 import com.main19.server.follow.mapper.FollowMapper;
 import com.main19.server.follow.service.FollowService;
