+++ server/src/main/java/com/main19/server/posting/controller/PostingController.java
@@ -1,6 +1,6 @@
 package com.main19.server.posting.controller;
 
-import com.main19.server.auth.Login;
+import com.main19.server.utils.Login;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -20,7 +20,6 @@ import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RequestPart;
@@ -29,8 +28,6 @@ import org.springframework.web.multipart.MultipartFile;
 
 import com.main19.server.dto.MultiResponseDto;
 import com.main19.server.dto.SingleResponseDto;
-import com.main19.server.exception.BusinessLogicException;
-import com.main19.server.exception.ExceptionCode;
 import com.main19.server.posting.dto.PostingPatchDto;
 import com.main19.server.posting.dto.PostingPostDto;
 import com.main19.server.posting.entity.Posting;
