+++ server/src/main/java/com/main19/server/posting/like/controller/PostingLikeController.java
@@ -1,6 +1,6 @@
 package com.main19.server.posting.like.controller;
 
-import com.main19.server.auth.Login;
+import com.main19.server.utils.Login;
 import com.main19.server.member.entity.Member;
 import javax.validation.Valid;
 import javax.validation.constraints.Positive;
@@ -12,7 +12,6 @@ import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
