+++ server/src/test/java/com/main19/server/comment/CommentControllerRestDocs.java
@@ -19,7 +19,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.google.gson.Gson;
-import com.main19.server.auth.LoginUserArgumentResolver;
+import com.main19.server.utils.LoginUserArgumentResolver;
 import com.main19.server.comment.controller.CommentController;
 import com.main19.server.comment.dto.CommentDto;
 import com.main19.server.comment.dto.CommentDto.Patch;
