+++ server/src/test/java/com/main19/server/comment/CommentLikeControllerRestDocs.java
@@ -18,7 +18,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.google.gson.Gson;
-import com.main19.server.auth.LoginUserArgumentResolver;
+import com.main19.server.utils.LoginUserArgumentResolver;
 import com.main19.server.comment.like.controller.CommentLikeController;
 import com.main19.server.comment.like.dto.CommentLikeDto;
 import com.main19.server.comment.like.dto.CommentLikeDto.Post;
@@ -26,11 +26,6 @@ import com.main19.server.comment.like.dto.CommentLikeDto.Response;
 import com.main19.server.comment.like.entity.CommentLike;
 import com.main19.server.comment.like.mapper.CommentLikeMapper;
 import com.main19.server.comment.like.service.CommentLikeService;
-import com.main19.server.posting.like.dto.PostingLikeDto;
-import com.main19.server.posting.like.dto.PostingLikeResponseDto;
-import com.main19.server.posting.like.entity.PostingLike;
-import com.main19.server.posting.like.service.PostingLikeService;
-import com.main19.server.posting.mapper.PostingMapper;
 import java.util.List;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
