+++ server/src/test/java/com/main19/server/gallery/GalleryControllerRestDocs.java
@@ -18,7 +18,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.google.gson.Gson;
-import com.main19.server.auth.LoginUserArgumentResolver;
+import com.main19.server.utils.LoginUserArgumentResolver;
 import com.main19.server.myplants.entity.MyPlants;
 import com.main19.server.myplants.gallery.controller.GalleryController;
 import com.main19.server.myplants.gallery.dto.GalleryDto;
