+++ server/src/test/java/com/main19/server/sse/SseRestDocs.java
@@ -17,8 +17,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import com.google.gson.Gson;
-import com.main19.server.auth.LoginUserArgumentResolver;
-import com.main19.server.config.WebConfig;
+import com.main19.server.utils.LoginUserArgumentResolver;
 import com.main19.server.member.entity.Member;
 import com.main19.server.posting.entity.Posting;
 import com.main19.server.sse.controller.SseController;
